Extend issuer check to also check signatures
The "This cert is not the issuer of the previous cert" check was
originally made just to highlight chains presented in the wrong order,
and only checked that the named matched. But now that there's a nice
handy way to do a real issuer check with cryptography we can do that.
This should also highlight when there's several CA certs with the same
name (but different keys) and the wrong one is in use.