Handle shortlived certs
Shortlived cert may not live a full month, so let's not mark them as
"will expire soon" for their entire lifetime. Use the same logic as
certbot and Let's Encrypt does with 1/2 for short lived certs, and 1/3
for "normal certs", but fall back to 30 days for certs with longer
validity than 90 days.